+++ app/Modules/Employees/Models/Employee.php
@@ -12,6 +12,7 @@ namespace BT\Modules\Employees\Models;
 
 use BT\Modules\Scheduler\Models\ScheduleResource;
 use BT\Support\CurrencyFormatter;
+use BT\Support\DateFormatter;
 use Illuminate\Database\Eloquent\Model;
 
 class Employee extends Model
@@ -38,6 +39,11 @@ class Employee extends Model
             ->where('resource_table','=','employees');
     }
 
+    public function type()
+    {
+        return $this->belongsTo('BT\Modules\Employees\Models\EmployeeType');
+    }
+
     //mutators
     public function setFirstNameAttribute($value){
         $this->attributes['first_name'] = ucfirst($value);
@@ -54,11 +60,31 @@ class Employee extends Model
     public function setShortNameAttribute($value){
         $this->attributes['short_name'] = $this->attributes['first_name'] . ' ' . substr($this->attributes['last_name'],0,1) . '.';
     }
-
+    //getters
     public function getFormattedBillingRateAttribute()
     {
         return CurrencyFormatter::format($this->attributes['billing_rate']);
     }
+
+    public function getFormattedTermDateAttribute()
+    {
+        if (!is_null($this->attributes['term_date'])) {
+            return DateFormatter::format($this->attributes['term_date']);
+        }
+    }
+
+    public function getFormattedScheduleAttribute(){
+        return $this->schedule ? trans('bt.yes') : trans('bt.no');
+    }
+
+    public function getFormattedActiveAttribute(){
+        return $this->active ? trans('bt.yes') : trans('bt.no');
+    }
+
+    public function getFormattedDriverAttribute(){
+        return $this->driver ? trans('bt.yes') : trans('bt.no');
+    }
+
 
     public function scopeStatus($query, $status)
     {