+++ app/Modules/Employees/Models/EmployeeType.php
@@ -1,0 +1,18 @@
+<?php
+
+namespace BT\Modules\Employees\Models;
+
+use Eloquent;
+
+/**
+ * Class Title
+ */
+class EmployeeType extends Eloquent
+{
+    /**
+     * @var bool
+     */
+    public $timestamps = false;
+
+
+}