+++ app/Modules/Documentation/Views/documentation/Release-Notes.md
@@ -1,5 +1,9 @@
 Release Notes
 ---
+### BillingTrack 6.0.3 (2022-07-01)
+- add employee type and termination date
+- update employees available query to include null term_date or term_date > $date
+- fix datatable status on paginate
 ### BillingTrack 6.0.2 (2022-06-14)
 - fix workorderToInvoice date setting
 ### BillingTrack 6.0.1 (2022-05-01)